[skip ci] [WIP] Move tuners to subpackages
